fix: resolve Next.js route export validation error
- Move sendProgress and session management to separate migration-sse-utils
- Remove invalid exports from Next.js route handlers
- Update imports to use new utility location
- Ensure only GET handler is exported from route files

Fixes Next.js type error about invalid route export fields.
